+++ rc_visard_driver/src/publishers/points2_publisher.h
@@ -58,9 +58,12 @@ public:
 
   Points2Publisher(ros::NodeHandle& nh, const std::string& frame_id_prefix, double f, double t, double scale);
 
+  void setOut1Alternate(bool alternate);
+
   bool used() override;
 
-  void publish(const rcg::Buffer* buffer, uint64_t pixelformat) override;
+  void publish(const rcg::Buffer* buffer, uint32_t part, uint64_t pixelformat) override;
+  void publish(const rcg::Buffer* buffer, uint32_t part, uint64_t pixelformat, bool out1);
 
 private:
   Points2Publisher(const Points2Publisher&);             // forbidden
@@ -73,6 +76,8 @@ private:
   float f;
   float t;
   float scale;
+
+  uint64_t tolerance_ns;
 
   ros::Publisher pub;
 };